refactor(common): minor NgOptimizedImage directive updates
This commit contains a set of small changes based on the feedback received in #47082.
