+++ packages/core/test/bundling/image-directive/e2e/basic/basic.e2e-spec.ts
@@ -9,7 +9,7 @@
 import {browser, by, element} from 'protractor';
 import {logging} from 'selenium-webdriver';
 
-import {collectBrowserLogs} from '../util';
+import {collectBrowserLogs} from '../browser-logs-util';
 
 describe('NgOptimizedImage directive', () => {
   it('should render an image with an updated `src`', async () => {
