+++ packages/core/test/bundling/image-directive/e2e/browser-logs-util.ts
