+++ packages/core/test/bundling/image-directive/e2e/preconnect-check/preconnect-check.e2e-spec.ts
@@ -10,7 +10,7 @@
 import {browser, by, element, ElementHelper} from 'protractor';
 import {logging} from 'selenium-webdriver';
 
-import {collectBrowserLogs} from '../util';
+import {collectBrowserLogs} from '../browser-logs-util';
 
 // Verifies that both images used in a component were rendered.
 async function verifyImagesPresent(element: ElementHelper) {
